# Reset connection asynchronously on return to pool