This PR implements asynchronous resetting of the connection when it's returned to the pool. This implements #178 and obsoletes PR #179.

## Important Note

The major side-effect of this PR (apart from increased performance) is that failures in `ResetConnectionAsync` will be silently ignored. This will suppress #169 and #208; the only side-effect will (ironically) be decreased performance if connections are getting destroyed when they're closed, instead of returned to the pool. I can't think of a good way of logging/diagnosing this (unless there's a logging/tracing framework in `netstandard1.3`?). (We could add a static `MySqlConnection.OnError` event handler or similar that clients could attach a logging callback to, but it would be nice to have something less manual and non-standard.)

## Benchmarks

### Single Threaded

The primary speed up is in single-threaded (low-concurrency) operation. Here are results from "Benchmark" when connecting to a LAN DB server 0.4ms away:

#### Before

Method | Mean | StdDev | StdErr | Min | Q1 | Median | Q3 | Max | Op/s | Allocated |
---------------------- |------------ |----------- |---------- |------------ |------------ |------------ |------------ |------------ |-------- |---------- |
OpenFromPoolNewAsync | 800.6765 us | 21.0694 us | 5.4401 us | 762.2547 us | 782.9775 us | 799.6714 us | 810.2822 us | 836.4873 us | 1248.94 | 16.97 kB |
OpenFromPoolNewSync | 722.4660 us | 24.6377 us | 6.3614 us | 671.8728 us | 706.7121 us | 724.8340 us | 736.4763 us | 765.1436 us | 1384.15 | 4.45 kB |

#### After

Method | Mean | StdErr | StdDev | Min | Q1 | Median | Q3 | Max | Op/s | Allocated |
---------------------- |------------ |---------- |----------- |------------ |------------ |------------ |------------ |------------ |-------- |---------- |
OpenFromPoolNewAsync | 346.8272 us | 3.7775 us | 14.1340 us | 328.4828 us | 338.2204 us | 345.1184 us | 349.9636 us | 382.3111 us | 2883.28 | 16.9 kB |
OpenFromPoolNewSync | 308.6513 us | 3.0141 us | 11.6738 us | 296.0878 us | 299.0418 us | 302.9528 us | 317.7077 us | 332.5841 us | 3239.9 | 12.24 kB |

### Concurrent

I ran the `MySqlConnector.Performance` tests under two configurations: a maximum pool size of 500 (so blocking on the connection pool should happen) and a maximum pool size of 2000 (should be no blocking). All tests were run against a local `mysqld` server. The command was `.\stress.ps1 N 30s async`.

##### Summary

When | MaxPoolSize | Requests | Mean | 50th | 95th | 99th | Max
--- | --- | --- | --- | --- | --- | --- | ---
Before | 500 | 1000 | 1.380799ms | 1.0042ms | 2.7962ms | 2.9843ms | 27.9794ms
After | 500 | 1000 | 1.106233ms | 997.7µs | 2.5788ms | 2.9801ms | 32.0018ms
Before | 500 | 2000 | 2.763622ms | 1.0009ms | 6.8465ms | 32.5073ms | 132.5081ms
After | 500 | 2000 | 2.050698ms | 999.9µs | 5.919ms | 18.5076ms | 87.9365ms
Before | 500 | 3000 | 4.332318ms | 1.0009ms | 17.4993ms | 58.9991ms | 206.0083ms
After | 500 | 3000 | 2.534541ms | 1.0003ms | 7.0002ms | 20.0056ms | 122.0119ms
Before | 2000 | 1000 | 1.371715ms | 1.0038ms | 2.785ms | 2.9723ms | 44.5096ms
After | 2000 | 1000 | 1.143765ms | 998.9µs | 2.6757ms | 2.9806ms | 52.8817ms
Before | 2000 | 2000 | 2.781073ms | 1.0012ms | 6.5657ms | 31.0009ms | 189.8881ms
After | 2000 | 2000 | 2.131828ms | 1.0002ms | 5.9785ms | 20.0009ms | 134.6819ms
Before | 2000 | 3000 | 4.856396ms | 1.0014ms | 23.0035ms | 60.502ms | 162.5065ms
After | 2000 | 3000 | 3.554666ms | 1.0012ms | 11.9998ms | 41.0099ms | 137.0207ms

#### Before

##### Maximum Pool Size = 500

```
Rate = 1000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            30000, 1000.03
Duration      [total, attack, wait]    30.0017319s, 29.999s, 2.7319ms
Latencies     [mean, 50, 95, 99, max]  1.380799ms, 1.0042ms, 2.7962ms, 2.9843ms, 27.9794ms
Bytes In      [total, mean]            346308890, 11543.63
Bytes Out     [total, mean]            11122500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:30000
Error Set:

Rate = 2000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            60000, 2000.03
Duration      [total, attack, wait]    30.0024226s, 29.9995s, 2.9226ms
Latencies     [mean, 50, 95, 99, max]  2.763622ms, 1.0009ms, 6.8465ms, 32.5073ms, 132.5081ms
Bytes In      [total, mean]            693267568, 11554.46
Bytes Out     [total, mean]            22245000, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:60000
Error Set:

Rate = 3000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            90000, 3000.00
Duration      [total, attack, wait]    30.001002s, 30.0000023s, 999.7µs
Latencies     [mean, 50, 95, 99, max]  4.332318ms, 1.0009ms, 17.4993ms, 58.9991ms, 206.0083ms
Bytes In      [total, mean]            1040017500, 11555.75
Bytes Out     [total, mean]            33367500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:90000
Error Set:
```

##### Maximum Pool Size = 2000

```
Rate = 1000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            30000, 1000.00
Duration      [total, attack, wait]    29.9999892s, 29.9999892s, 0s
Latencies     [mean, 50, 95, 99, max]  1.371715ms, 1.0038ms, 2.785ms, 2.9723ms, 44.5096ms
Bytes In      [total, mean]            346595080, 11553.17
Bytes Out     [total, mean]            11122500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:30000
Error Set:

Rate = 2000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            60000, 2000.03
Duration      [total, attack, wait]    30.0041588s, 29.9995s, 4.6588ms
Latencies     [mean, 50, 95, 99, max]  2.781073ms, 1.0012ms, 6.5657ms, 31.0009ms, 189.8881ms
Bytes In      [total, mean]            693345000, 11555.75
Bytes Out     [total, mean]            22245000, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:60000
Error Set:

Rate = 3000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            90000, 3000.01
Duration      [total, attack, wait]    30.0048949s, 29.9998982s, 4.9967ms
Latencies     [mean, 50, 95, 99, max]  4.856396ms, 1.0014ms, 23.0035ms, 60.502ms, 162.5065ms
Bytes In      [total, mean]            1040017500, 11555.75
Bytes Out     [total, mean]            33367500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:90000
Error Set:
```

#### After

##### Maximum Pool Size = 500

```
Rate = 1000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            30000, 1000.01
Duration      [total, attack, wait]    29.9995624s, 29.9995624s, 0s
Latencies     [mean, 50, 95, 99, max]  1.106233ms, 997.7µs, 2.5788ms, 2.9801ms, 32.0018ms
Bytes In      [total, mean]            346307402, 11543.58
Bytes Out     [total, mean]            11122500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:30000
Error Set:

Rate = 2000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            60000, 2000.00
Duration      [total, attack, wait]    30.0019398s, 29.9999408s, 1.999ms
Latencies     [mean, 50, 95, 99, max]  2.050698ms, 999.9µs, 5.919ms, 18.5076ms, 87.9365ms
Bytes In      [total, mean]            693267566, 11554.46
Bytes Out     [total, mean]            22245000, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:60000
Error Set:

Rate = 3000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            90000, 3000.04
Duration      [total, attack, wait]    30.001378s, 29.999636667s, 1.741333ms
Latencies     [mean, 50, 95, 99, max]  2.534541ms, 1.0003ms, 7.0002ms, 20.0056ms, 122.0119ms
Bytes In      [total, mean]            1040017500, 11555.75
Bytes Out     [total, mean]            33367500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:90000
Error Set:
```

##### Maximum Pool Size = 2000

```
Rate = 1000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            30000, 1000.01
Duration      [total, attack, wait]    29.9998073s, 29.9998073s, 0s
Latencies     [mean, 50, 95, 99, max]  1.143765ms, 998.9µs, 2.6757ms, 2.9806ms, 52.8817ms
Bytes In      [total, mean]            346304431, 11543.48
Bytes Out     [total, mean]            11122500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:30000
Error Set:

Rate = 2000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            60000, 1999.95
Duration      [total, attack, wait]    30.0007196s, 30.0007196s, 0s
Latencies     [mean, 50, 95, 99, max]  2.131828ms, 1.0002ms, 5.9785ms, 20.0009ms, 134.6819ms
Bytes In      [total, mean]            693267576, 11554.46
Bytes Out     [total, mean]            22245000, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:60000
Error Set:

Rate = 3000 RPS
Duration = 30s
Targets = targets-async.txt
Requests      [total, rate]            90000, 3000.01
Duration      [total, attack, wait]    30.0033768s, 29.9998718s, 3.505ms
Latencies     [mean, 50, 95, 99, max]  3.554666ms, 1.0012ms, 11.9998ms, 41.0099ms, 137.0207ms
Bytes In      [total, mean]            1040017500, 11555.75
Bytes Out     [total, mean]            33367500, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:90000
Error Set:
```