@caleblloyd 2017-05-16
This approach is causing thread pool lockups on sync codepaths.  Sync Performance Tests from `master`:

```
$ ./stress.sh 1000 1s sync
Rate = 1000 RPS
Duration = 1s
Targets = targets-sync.txt
Requests      [total, rate]            1000, 1001.00
Duration      [total, attack, wait]    1.002371974s, 998.99968ms, 3.372294ms
Latencies     [mean, 50, 95, 99, max]  63.843849ms, 62.906768ms, 112.59616ms, 153.395002ms, 166.684714ms
Bytes In      [total, mean]            11349417, 11349.42
Bytes Out     [total, mean]            370750, 370.75
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:1000  
```

Sync performance tests with this PR (after a fresh start, before the thread pool has had time to warm up).  I had to stop the test after about 30 seconds because everything was locked up, but as you can see there's issues:

```
./stress.sh 1000 1s sync
Rate = 1000 RPS
Duration = 1s
Targets = targets-sync.txt
Requests      [total, rate]            998, 999.00
Duration      [total, attack, wait]    30.999888938s, 998.999759ms, 30.000889179s
Latencies     [mean, 50, 95, 99, max]  21.845362043s, 30.000623168s, 30.003780265s, 30.017442424s, 30.026976038s
Bytes In      [total, mean]            3085915, 3092.10
Bytes Out     [total, mean]            105293, 105.50
Success       [ratio]                  27.25%
Status Codes  [code:count]             200:272  500:6  0:720  
```

I think that for this approach to work, we'd need to shift to a "hybrid approach", where connections can either get reset upon return in an async manner, or they can get reset when drawn.  Queue the connection but don't mark it as reset.  Then kick off the async reset task in `Dispose`.  If the connection is drawn from the pool again and still has not been reset, perform the reset there and cancel the async reset task.  Something along those lines.