@bgrainger 2017-05-16
Sigh, I assume this is the "classic" problem of async I/O (or any queued threadpool work) being backlogged when the threadpool is flooded with blocking requests.

I'm going to close this for now; it might be something to revisit in `netstandard2.0` when we could use `Thread` to create a dedicated connection resetting thread. In the meantime, it could be worth tackling #258 and remove `TryPingAsync` to try to reduce connection opening latency.