@caleblloyd 2017-05-17
> Sigh, I assume this is the "classic" problem of async I/O (or any queued threadpool work) being backlogged when the threadpool is flooded with blocking requests.

Yes, the same lock-up behavior can be observed with the code today if you mix a bunch of Sync and Async requests together. It can be observed when combining the `scripts/vegeta/targets-sync.txt` and `scripts/vegeta/targets-async.txt` files into a `targets-both.txt` file and run `./stress.sh 10000 1s both`.

I'm curious as to what the official stance is here.  I want to say it's up to the caller to schedule Async work and Sync work onto the proper thread pools.  I think that since we provide truly Async and truly Sync methods, we have done our part and shouldn't be involved in trying to sort out what threads code runs on.

> It might be something to revisit in `netstandard2.0` when we could use Thread to create a dedicated connection resetting thread.

Steven Toub's [comment](https://github.com/dotnet/corefx/issues/2576#issuecomment-126693306) mentions that "by passing TaskCreationOptions.LongRunning and TaskScheduler.Default to StartNew, a dedicated thread will be created for that one Task instance", so this may be possible without `netstandard2.0`.

Since this is a library, I have reservations about creating threads, thread pools, buffer pools, etc.  I think that we _can_ do it so long as we expose configuration options to the caller and choose sensible defaults.